Upgrade to CEF 105.3.18

+++ AvaloniaApp/CustomWebViewGlue.cs
@@ -42,7 +42,7 @@ namespace AvaloniaApp
 			CefMenuModel submenu = model.AddSubMenu(0, "Submenu Test");
 			submenu.AddItem((int)CefMenuId.Copy, "Copy");
 			submenu.AddItem((int)CefMenuId.Paste, "Paste");
-			submenu.SetColorAt(submenu.Count - 1, CefMenuColorType.Text, CefColor.FromArgb((int)Colors.Blue.ToUint32()));
+			submenu.SetColorAt((int)submenu.Count - 1, CefMenuColorType.Text, CefColor.FromArgb((int)Colors.Blue.ToUint32()));
 			submenu.AddCheckItem(0, "Checked Test");
 			submenu.SetCheckedAt(submenu.Count - 1, true);
 			submenu.AddRadioItem(0, "Radio Off", 0);
