+++ CefNet.Modern.Forms/Internal/ModernFormsContextMenuRunner.cs
@@ -136,7 +136,7 @@ namespace CefNet.Internal
 		private void Build(CefMenuModel model, MenuItemCollection menu)
 		{
 			CefColor color = default;
-			int count = model.Count;
+			long count = model.Count;
 			for (int i = 0; i < count; i++)
 			{
 				MenuItem menuItem;
